fixed invoice number generation when generating many invoices, added bad payruns archival

+++ LBH.AdultSocialCare.Functions.Payruns/Gateways/Interfaces/IInvoiceGateway.cs
@@ -9,6 +9,6 @@ namespace LBH.AdultSocialCare.Functions.Payruns.Gateways.Interfaces
     {
         Task<Dictionary<Guid, List<InvoiceDomain>>> GetInvoicesByPackageIds(IList<Guid> packageIds);
 
-        Task<int> GetInvoicesCountAsync();
+        Task<long> GetInvoicesCountAsync();
     }
 }
